Fix pylint linting errors
Refactor code to address linting issues:

Backend Changes:
- Remove unused imports from user_resources.py (ForeignKey, relationship)
- Refactor main.py to reduce line count from 1085 to 830 lines
  - Extract user resources endpoints to separate router
  - Create new api/routes structure
  - Move all user resources endpoints to api/routes/user_resources.py
  - Include router in main.py using app.include_router()

This refactoring improves code organization and maintainability while
resolving all pylint errors:
✅ Fixed C0302: Too many lines in module (830/1000)
✅ Fixed W0611: Unused imports
