+++ backend/app/api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""API routes package initialization."""
