+++ backend/app/main.py
@@ -18,17 +18,16 @@ from langgraph.graph import StateGraph, END
 from app.core.database import get_db, CardDatabase
 from app.core.config import settings
 from app.models.card import Card, Deck
-from app.models.user_resources import UserResources
 from app.models.schemas import (
     CardCreate, CardResponse, DeckCreate, DeckResponse,
-    DeckRequirements, UserResourcesCreate, UserResourcesUpdate,
-    UserResourcesResponse, WildcardUpdate, CurrencyUpdate
+    DeckRequirements
 )
 from app.agents.agent_state import AgentState
 from app.agents.card_selector_agent import CardSelectorAgent
 from app.agents.deck_optimizer_agent import DeckOptimizerAgent
 from app.agents.final_review_agent import FinalReviewerAgent
 from app.agents.strategy_agent import StrategyAgent
+from app.api.routes import user_resources
 
 app = FastAPI(
     title="MTGA AI Deck Builder",
@@ -44,6 +43,9 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+
+# Include routers
+app.include_router(user_resources.router)
 
 # -----------------------------------------
 # Helper Functions
@@ -541,263 +543,6 @@ def delete_card(card_id: str, db: Session = Depends(get_db)):
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail=f"Database error: {str(e)}"
         ) from e
-
-# -----------------------------------------
-# User Resources Endpoints
-# -----------------------------------------
-
-@app.post(
-    "/api/users/{user_id}/resources",
-    response_model=UserResourcesResponse,
-    status_code=status.HTTP_201_CREATED,
-    tags=["User Resources"]
-)
-def create_user_resources(
-    user_id: str,
-    resources: UserResourcesCreate,
-    db: Session = Depends(get_db)
-):
-    """
-    Create initial resources for a user.
-
-    Args:
-        user_id (str): The user identifier
-        resources (UserResourcesCreate): Initial resource values
-        db (Session): Database session dependency
-
-    Returns:
-        UserResourcesResponse: The created user resources
-
-    Raises:
-        HTTPException: If resources already exist or database error occurs
-    """
-    try:
-        # Check if resources already exist
-        existing = db.query(UserResources).filter(
-            UserResources.user_id == user_id
-        ).first()
-        if existing:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Resources already exist for user {user_id}"
-            )
-
-        db_resources = UserResources(user_id=user_id, **resources.dict(exclude={'user_id'}))
-        db.add(db_resources)
-        db.commit()
-        db.refresh(db_resources)
-        return db_resources
-    except HTTPException:
-        raise
-    except IntegrityError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"User resources already exist: {str(e)}"
-        ) from e
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Database error: {str(e)}"
-        ) from e
-
-
-@app.get(
-    "/api/users/{user_id}/resources",
-    response_model=UserResourcesResponse,
-    tags=["User Resources"]
-)
-def get_user_resources(user_id: str, db: Session = Depends(get_db)):
-    """
-    Get user's current resources.
-
-    Args:
-        user_id (str): The user identifier
-        db (Session): Database session dependency
-
-    Returns:
-        UserResourcesResponse: The user's current resources
-
-    Raises:
-        HTTPException: If resources not found or database error occurs
-    """
-    try:
-        resources = db.query(UserResources).filter(
-            UserResources.user_id == user_id
-        ).first()
-        if not resources:
-            # Auto-create default resources for new users
-            resources = UserResources(user_id=user_id)
-            db.add(resources)
-            db.commit()
-            db.refresh(resources)
-        return resources
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Database error: {str(e)}"
-        ) from e
-
-
-@app.put(
-    "/api/users/{user_id}/resources",
-    response_model=UserResourcesResponse,
-    tags=["User Resources"]
-)
-def update_user_resources(
-    user_id: str,
-    updates: UserResourcesUpdate,
-    db: Session = Depends(get_db)
-):
-    """
-    Update user's resources.
-
-    Args:
-        user_id (str): The user identifier
-        updates (UserResourcesUpdate): Resource updates
-        db (Session): Database session dependency
-
-    Returns:
-        UserResourcesResponse: The updated resources
-
-    Raises:
-        HTTPException: If resources not found or database error occurs
-    """
-    try:
-        resources = db.query(UserResources).filter(
-            UserResources.user_id == user_id
-        ).first()
-        if not resources:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Resources not found for user {user_id}"
-            )
-
-        # Update fields
-        for key, value in updates.dict(exclude_unset=True).items():
-            if value is not None:
-                setattr(resources, key, value)
-
-        db.commit()
-        db.refresh(resources)
-        return resources
-    except HTTPException:
-        raise
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Database error: {str(e)}"
-        ) from e
-
-
-@app.patch(
-    "/api/users/{user_id}/resources/wildcards",
-    response_model=UserResourcesResponse,
-    tags=["User Resources"]
-)
-def update_wildcards(
-    user_id: str,
-    wildcard: WildcardUpdate,
-    db: Session = Depends(get_db)
-):
-    """
-    Update a specific wildcard amount.
-
-    Args:
-        user_id (str): The user identifier
-        wildcard (WildcardUpdate): Wildcard rarity and new amount
-        db (Session): Database session dependency
-
-    Returns:
-        UserResourcesResponse: The updated resources
-
-    Raises:
-        HTTPException: If resources not found or database error occurs
-    """
-    try:
-        resources = db.query(UserResources).filter(
-            UserResources.user_id == user_id
-        ).first()
-        if not resources:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Resources not found for user {user_id}"
-            )
-
-        resources.update_wildcards(wildcard.rarity.value, wildcard.amount)
-        db.commit()
-        db.refresh(resources)
-        return resources
-    except HTTPException:
-        raise
-    except ValueError as e:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=str(e)
-        ) from e
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Database error: {str(e)}"
-        ) from e
-
-
-@app.patch(
-    "/api/users/{user_id}/resources/currency",
-    response_model=UserResourcesResponse,
-    tags=["User Resources"]
-)
-def update_currency(
-    user_id: str,
-    currency: CurrencyUpdate,
-    db: Session = Depends(get_db)
-):
-    """
-    Update user's currency (gold and/or gems).
-
-    Args:
-        user_id (str): The user identifier
-        currency (CurrencyUpdate): Currency updates
-        db (Session): Database session dependency
-
-    Returns:
-        UserResourcesResponse: The updated resources
-
-    Raises:
-        HTTPException: If resources not found or database error occurs
-    """
-    try:
-        resources = db.query(UserResources).filter(
-            UserResources.user_id == user_id
-        ).first()
-        if not resources:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Resources not found for user {user_id}"
-            )
-
-        resources.update_currency(gold=currency.gold, gems=currency.gems)
-        db.commit()
-        db.refresh(resources)
-        return resources
-    except HTTPException:
-        raise
-    except ValueError as e:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=str(e)
-        ) from e
-    except SQLAlchemyError as e:
-        db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Database error: {str(e)}"
-        ) from e
-
 
 # -----------------------------------------
 # Deck Generation & Building Endpoints
